+++ packages/components/quotation/index.js
@@ -1,3 +1,2 @@
 export * from './modals/index.js';
 export * from './views/index.js';
-export { mountQuotationFlow } from './logic/createQuotationFlowComponent.js';